sua loi hien thi anh

+++ backend/src/main/java/com/techroom/roommanagement/controller/AdminController.java
@@ -41,7 +41,7 @@ public class AdminController {
     @PutMapping("/users/{userId}/role")
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<?> updateUserRole(
-            @PathVariable Long userId,
+            @PathVariable int userId,
             @RequestParam int newRole) {
         try {
             User user = userRepository.findById(userId)
