+++ backend/src/main/java/com/techroom/roommanagement/model/RefreshToken.java
@@ -11,13 +11,13 @@ public class RefreshToken {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(nullable = false, unique = true, length = 500)
     private String token;
 
     @Column(name = "user_id", nullable = false)
-    private Long userId;
+    private int userId;
 
     @Column(name = "expiry_date", nullable = false)
     private LocalDateTime expiryDate;
