+++ backend/src/main/java/com/techroom/roommanagement/repository/RefreshTokenRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
     Optional<RefreshToken> findByToken(String token);
-    void deleteByUserId(Long userId);
+    void deleteByUserId(Integer userId);
     void deleteByToken(String token);
 }
